src/requirements.txt
 python-dotenv
 python-telegram-bot